pyproject.toml
[project]
name = "mask"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.3",
    "django-pytest>=0.2.0",
    "django-rest-framework>=0.1.0",
    "pandas>=2.3.0",
    "pytest>=8.4.1",
]

[dependency-groups]
lint = [
    "ruff>=0.12.1",
]
